Add files via upload
Exploratory Data Analysis has been done on the sales data of a restaurant. I'd like to point out that post-analysis necessary inferences have been duly observed and recorded. The EDA has helped in figuring out the past trends and necessary suggestions can be made for optimized growth of the restaurant.